Fix final batch during an epoch by padding it.
Finally gonna be seperating branches yay! (and more professional commit
messages xD)
Mainly because from the last commit (01bcdc6), the training differed too
much between the devices I had.
For example, the costs exploded when trying to train on some gpu (to
ridiculous numbers like 10000~ or the max for cce after
'converging').
Whereas, on an nvidia gpu, it converged at relatively high numbers
(~1).
And, on my intel gpu, a lower convergeance but still much higher than
without the last batch.
I tried to fix that in this commit by adding padding which is what I
should have done from the start.

Will have to see if costs still explode on that first gpu and see if
padding wasn't the problem which I don't think it is...