Contains atomic averaging code generation commented out
Entirely changed the gpu memory layout so it doesn't require
array<GradWeights, batch_size> and just GradWeights (this could be
changed to atomic<u32> in workgroup memory so it doesn't require a
seperate Struct though I'm not sure)

While doing both accumalation and the averaging on the gpu feels a lot
more natural, it also feels weirder. Hence, I'm choosing the option of
division on the cpu. I wish for someone else to performance test this
since I really can't be bothered by it and so, I'll be removing the
atomic averaging code generation from now on.